pyproject.toml
[project]
name = "bluedot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.78.0",
    "asyncio>=4.0.0",
    "bitsandbytes>=0.49.1",
    "python-dotenv>=1.0.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "tqdm>=4.60.0",
    "transformers>=4.57.6",
]
